wt-image: Tell borg not to complain about unencrypted repositories.
If not explcitly told so it asks whether it is ok to access it which is
very confusing since most of its output is redirected, causing it to
eventually fail in a cryptic way.
By their nature these Windows images are not expected to contain
confidential data and thus shouldn't require an encrypted repository so
setting BORG_UNKNOWN_UNENCRYPTED_REPO_ACCESS_IS_OK should be fine.

Signed-off-by: Francois Gouget <fgouget@free.fr>