wt-install-dev: Rename --ignore to --test in deb_{pkg,multi}().
This option is used to test whether a package exists / is multiarch for
the purpose of installing another package, hence the new name.
As before, when given this option deb_pkg() and deb_multi() will not add
the package to the missing / broken package lists but will still
schedule its installation as it may provide header files and libraries
needed by the workarounds for the main package.

Signed-off-by: Francois Gouget <fgouget@free.fr>