wt-install-dev: Always add broken Debian packages to the list.
The pattern to deal with Debian packages where multiarch support is
broken by dependencies was to test the dependencies and only call
deb_multi() if the test showed it was ok. This means the package would
not be added to the broken packages list. It also forced the code to
separately install the 64-bit version and the matching library package.
deb_multi() can now take the exit code from the test with the --if-0 and
--if-not-0 options and use that to replace the multiarch compatibility
test. This makes it possible to always call deb_multi(), simplifying the
code and ensuring no broken package is missing from the list.

Signed-off-by: Francois Gouget <fgouget@free.fr>