wt-install-dev: Separate initialisation from package installation scheduling.
Move the calls to the platform-specific functions that schedule packages
for installation from wt_init() to wt_pre_install().
This allows scripts that use wt-install-dev as a library to inspect the
result of wt_init() and to bail out if appropriate, such as if they
don't support the detected platform.

Signed-off-by: Francois Gouget <fgouget@free.fr>