wt-daily & wt-check: Queue the other OS tests and refresh the commit timestamps.
Without the queue if the first OS did not produce usable results
wt-check would delete the corresponding commit file, causing the tests
to be re-run in that first OS (mostly) before ever trying the others.
Now wt-check only cleans up the commit files once the queue is empty,
that is once the tests have been run at least once in each of the
other OSes.
Furthermore wt-daily refreshes the commit timestamp of commit files
once the test has completed, making it easier to ensure the website had
time to update.

Signed-off-by: Francois Gouget <fgouget@free.fr>