wt-install-dev: Properly handle non-existant zypper repositories.
This notably happens when running the script on an old release and the
repository only has files for the newer ones. Conversely it could happen
if the repository has not been updated to provide files for the latest
release.
In either case, make sure the addrepo command does not get stuck in the
'wt-install-dev --yes' case.

Signed-off-by: Francois Gouget <fgouget@free.fr>