wt-install-dev: Add GStreamer workarounds on openSUSE Leap and Tumbleweed.
gstreamer-devel-32bit fails to install a lot of its dependencies.
In particular a lot of 32-bit pkgconfig files are needed in order for
autoconf to detect that GStreamer is available. So add
zypper_create_pcs() to simplify creating symbolic links to work around
the missing files.
Also add zypper_is_installed() so the workarounds can be applied only
when it makes sense, thus avoiding spurious error messages.

Signed-off-by: Francois Gouget <fgouget@free.fr>